refactored configuration
